chore: move utils file back
